+++ CHANGELOG.md
@@ -14,3 +14,4 @@ Created on Dec 20, 2023
 - Initial Development: Account manager is completed on Decemeber 21, 2023 -> **0.22**
 - Initial Development: Implementing transaction system on Decemeber 22, 2023 -> **0.3**
 - Beta Release: Main functionalities such account management and transaction system can function on Decemeber 22, 2023 -> **0.4**
+- Beta Release: Added more option on Decemeber 22, 2023 -> **0.41**